+++ e2e/npm-ci-registry.js
@@ -108,7 +108,9 @@ EOD`;
    */
   publishComponent(componentName: string, componentVersion?: string = '0.0.1') {
     const packDir = path.join(this.helper.localScopePath, 'pack');
-    this.helper.runCmd(`bit npm-pack ${this.helper.remoteScope}/${componentName} -o -k -d ${packDir}`);
+    this.helper.runCmd(
+      `bit npm-pack ${this.helper.remoteScope}/${componentName}@${componentVersion} -o -k -d ${packDir}`
+    );
     const npmComponentName = componentName.replace(/\//g, '.');
     const tarballFileName = `ci-${this.helper.remoteScope}.${npmComponentName}-${componentVersion}.tgz`;
     const tarballFilePath = path.join(packDir, tarballFileName);
@@ -116,6 +118,27 @@ EOD`;
     const extractedDir = path.join(packDir, 'package');
     this._validateRegistryScope(extractedDir);
     this.helper.runCmd('npm publish', extractedDir);
+  }
+
+  /**
+   * a workaround to make Bit save dependencies as packages.
+   * once the resolver is set, it's possible to delete the remote and it'll be still able to import
+   * from that remote.
+   * once the remote scope is deleted from the remote list, Bit assumes that it the remote is a hub
+   * and enable the save-dependencies-as-packages feature.
+   */
+  setResolver() {
+    const scopeJsonPath = '.bit/scope.json';
+    const scopeJson = this.helper.readJsonFile(scopeJsonPath);
+    const resolverPath = path.join(this.helper.localScopePath, 'resolver.js');
+    // $FlowFixMe
+    scopeJson.resolverPath = resolverPath;
+    this.helper.createJsonFile(scopeJsonPath, scopeJson);
+    this.helper.createFile('', 'resolver.js', this._getResolverContent());
+  }
+
+  _getResolverContent() {
+    return `module.exports = () => Promise.resolve('file://${this.helper.remoteScopePath}');`;
   }
 
   _validateRegistryScope(dir: string) {