+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - fix `bit show --remote --json` to not crash when a component has a compiler
+- fix `bit checkout` to update bit.json with the checked out version
+- fix "Maximum call stack" error when resolving js files after css files (bit-javascript)
+- fix `bit checkout --all` to write the correct data when some components are also dependencies of others
+- fix `bit checkout` to install dependencies as packages when applicable
+- fix `bit remove --remote` to show the dependents correctly
 - hide component internal structure diff upon `bit diff` unless `--verbose` flag is used
 - implement postinstall symlink generation for cases when custom-resolve modules is used with unsupported file (such as binary files)
 - fix parsing `.tsx` files (bit-javascript)